pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cftable"
version = "0.1.0"
description = "Cash Flow Simulation Tool"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
cftable = "cftable.cli:run_cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["cftable*"]
